// types/vacancy.d.ts
export type Vacancy = {
    id: number,
    title: string,
    company: string,
    companyLogo: string,
    workSchedule: string,
    minSalary: number,
    maxSalary: number,
    location: string,
    education: string,
    jobLevel: string,
    experience: string,
    description: string,
    authorId: number,
    createdAt: string,
    updatedAt: string
}